requirements.txt
# Django Framework
Django>=4.2.0,<5.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Python

# Image Processing
Pillow>=10.0.0  # For handling image uploads (recipe photos, user avatars)

# Environment Variables
python-decouple>=3.8  # For managing .env files

# REST Framework (for API development)
djangorestframework>=3.14.0

# CORS Headers (for frontend integration)
django-cors-headers>=4.0.0

# Development Tools
ipython>=8.12.0  # Enhanced Python shell
django-extensions>=3.2.0  # Useful Django extensions

# Code Quality
black>=23.0.0  # Code formatter
flake8>=6.0.0  # Linter

# Security
django-environ>=0.11.0  # Environment variable management

